backend/requirements.txt
Django==5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.12
django-cors-headers==3.14.0
python-dotenv
cryptography
pyjwt==2.6.0
firebase-admin==6.0.1
pycryptodome==3.19.0
